refact: create types.ts
