+++ src/pizzaData.ts
@@ -1,19 +1,5 @@
-export type PizzaOptions = { value: string; name: string; price?: number };
-export type PizzaIngredients = {
-  value: string;
-  name: string;
-  url: string;
-  price: number;
-};
-
-interface PizzaData {
-  pizzaSize: PizzaOptions[];
-  pizzaDough: PizzaOptions[];
-  pizzaSauce: PizzaOptions[];
-  pizzaCheese: PizzaIngredients[];
-  pizzaVegetables: PizzaIngredients[];
-  pizzaMeat: PizzaIngredients[];
-}
+// Types
+import { PizzaOptions, PizzaIngredients, PizzaData } from './types';
 
 export const pizzaData: PizzaData = Object.freeze({
   pizzaSize: [
